% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_for_app.R
\name{compute_effectsize}
\alias{compute_effectsize}
\title{Compute effect size}
\usage{
compute_effectsize(raw_data_lg)
}
\arguments{
\item{raw_data_lg}{data.frame, cells values in long format}
}
\value{
data.frame, Cohen's effect size and fold change that were observed
in the data for each marker
}
\description{
Compute effect size
}
\details{
Compute observed Cohen's effect size and observed fold change
}
