% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_for_app.R
\name{compute_pwr}
\alias{compute_pwr}
\title{Compute power}
\usage{
compute_pwr(model_values, alpha = 0.05)
}
\arguments{
\item{model_values}{data.frame, output of run_models().}

\item{alpha}{numeric, significance level.}
}
\value{
numeric, power.
}
\description{
Compute power
}
\details{
Compute the power based on the model values.
}
\keyword{internal}
