% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_for_app.R
\name{compute_variance}
\alias{compute_variance}
\title{Compute variance}
\usage{
compute_variance(raw_data_lg)
}
\arguments{
\item{raw_data_lg}{data.frame, cells values in long format}
}
\value{
numeric, observed variance mean
}
\description{
Compute variance
}
\details{
Compute the observed variance in the data
}
