% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CyTOFpower.R, R/app.R
\docType{package}
\name{CyTOFpower}
\alias{CyTOFpower}
\title{CyTOFpower: R-package to predict the power of a CyTOF experiment}
\usage{
CyTOFpower()
}
\value{
Interactive shiny app.
}
\description{
This package provides functions and a shiny app to predict the power of a
CyTOF experiment.

Interactive shiny app to predict the power of a CyTOF experiment.
}
\examples{
# Launch the shiny app
if (interactive()) {
  CyTOFpower()
}
}
