% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models_onesim.R
\name{function_apply_modelcomputations_modelchoice}
\alias{function_apply_modelcomputations_modelchoice}
\title{Run DS tests for one simulation.}
\usage{
function_apply_modelcomputations_modelchoice(
  list_combined_output,
  model = c("cytoglmm", "cytoglm", "testDS_limma_random", "testDS_limma_fixed",
    "testDS_lmm")
)
}
\arguments{
\item{list_combined_output}{list, of simulated data (output of the
function_apply_onesimulation_withmarkerinfo function).}

\item{model}{vector, name(s) of models to test.}
}
\value{
data.frame of results for each simulation.
}
\description{
Run DS tests for one simulation.
}
\details{
Wrapper to run the models through the different simulations.
}
