% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_nb_parameters_optimized.R
\name{function_apply_onesimulation_withmarkerinfo}
\alias{function_apply_onesimulation_withmarkerinfo}
\title{One simulation with markers NB information.}
\usage{
function_apply_onesimulation_withmarkerinfo(variation)
}
\arguments{
\item{variation}{list, list of list containing the different input
parameter variations @describeIn function_create_mock_dataset_withmarkerinfo.}
}
\value{
list of simulated data. Each list member contains 5 slots:
    - variation: the variation of input paramaters which has been used in input;
    - df_info: experimental information (donor IDs, group IDs, samples IDs);
    - DE_markers_names: name of the differentially expressed markers;
    - ls_mock_data: list of data.frames, each data.frame being one simulation of
    the cell values using the input parameters provided (list is length nb_sim).
}
\description{
One simulation with markers NB information.
}
\details{
Apply one simulation for a given parameter combination
number times (iterations), when we have prior information about the markers
distribution parameters (mean and dispersion of the negative binomial).
}
\keyword{internal}
