% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_nb_parameters_optimized.R
\name{function_check_nbDEmarkers}
\alias{function_check_nbDEmarkers}
\title{Number of DE markers greater than 1.}
\usage{
function_check_nbDEmarkers(nb_DEmarker)
}
\arguments{
\item{nb_DEmarker}{numeric, number of differentially expressed markers.}
}
\value{
error message
}
\description{
Number of DE markers greater than 1.
}
\details{
Function to check that the number of DE markers is greater than 1.
}
