% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_nb_parameters_optimized.R
\name{function_check_nbmarkers}
\alias{function_check_nbmarkers}
\title{Number of markers greater than 2.}
\usage{
function_check_nbmarkers(nb_marker)
}
\arguments{
\item{nb_marker}{numeric, total number of markers.}
}
\value{
error message if the total number of markers is lower than 3.
}
\description{
Number of markers greater than 2.
}
\details{
Function to check that the number of markers is greater than 2.
}
\keyword{internal}
