% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{function_compute_diffcyt_features}
\alias{function_compute_diffcyt_features}
\title{Compute cells counts and medians.}
\usage{
function_compute_diffcyt_features(mock_flowset)
}
\arguments{
\item{mock_flowset}{data.frame, cell values for each marker.}
}
\value{
list with cell counts and medians for each markers.
}
\description{
Compute cells counts and medians.
}
\details{
Function to calculate features for diffcyt package models.
}
\keyword{internal}
