% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_nb_parameters_optimized.R
\name{function_DEmarkers_sup_nbmarkers}
\alias{function_DEmarkers_sup_nbmarkers}
\title{Check on the number of DE markers.}
\usage{
function_DEmarkers_sup_nbmarkers(nb_markers, nb_DEmarker)
}
\arguments{
\item{nb_markers}{numeric, total number of markers.}

\item{nb_DEmarker}{numeric, number of differentially expressed markers.}
}
\value{
error message if the number of DE markers is greater than the total
number of markers.
}
\description{
Check on the number of DE markers.
}
\details{
Function to check that number of DE markers greater than number of
markers.
}
