% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{function_desigmat_contrast_diffcytDSlimma_fixedeffect}
\alias{function_desigmat_contrast_diffcytDSlimma_fixedeffect}
\title{Design and contrast matrices for diffcyt-DS-limma with fixed effect.}
\usage{
function_desigmat_contrast_diffcytDSlimma_fixedeffect(df_experiment_info)
}
\arguments{
\item{df_experiment_info}{data.frame, information about the experiment (donor IDs,
group IDs and sample IDs).}
}
\value{
list with 3 slots:
    - design_matrix: the design matrix;
    - contrast: the contrast matrix;
    - effect: the specification of the "fixed" effect.
}
\description{
Design and contrast matrices for diffcyt-DS-limma with fixed effect.
}
\details{
Function to create the design matrix and contrast for the diffcyt-DS-limma model
with fixed effect.
}
