% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_nb_parameters_optimized.R
\name{function_extract_marker_names}
\alias{function_extract_marker_names}
\title{Extract marker names.}
\usage{
function_extract_marker_names(mock_dataset)
}
\arguments{
\item{mock_dataset}{data.frame, containing the cell values for each marker.}
}
\value{
vector of marker names.
}
\description{
Extract marker names.
}
\details{
Function to extract the marker names.
}
\keyword{internal}
