% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_nb_parameters_optimized.R
\name{function_is_data_paired}
\alias{function_is_data_paired}
\title{Data paired.}
\usage{
function_is_data_paired(data)
}
\arguments{
\item{data}{data.frame, experimental information containing (sample IDS,
donor IDs).}
}
\value{
logical, TRUE is the data are paired - FALSE if the data are not paired
}
\description{
Data paired.
}
\details{
Function to check if the data is paired.
}
\keyword{internal}
