% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_nb_parameters_optimized.R
\name{function_names_DE_markers}
\alias{function_names_DE_markers}
\title{Generate name of the DE markers.}
\usage{
function_names_DE_markers(total_nb_marker, nb_DE_marker)
}
\arguments{
\item{total_nb_marker}{numeric, total number of markers.}

\item{nb_DE_marker}{numeric, number of DE markers.}
}
\value{
vactor of marker names.
}
\description{
Generate name of the DE markers.
}
\details{
Function to generate the name of the DE markers.
}
\keyword{internal}
