% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{function_run_bootstrapcytoGLMM}
\alias{function_run_bootstrapcytoGLMM}
\title{Run cytoglm.}
\usage{
function_run_bootstrapcytoGLMM(mock_dataset, nb_bootstrap = 500)
}
\arguments{
\item{mock_dataset}{data.frame, cell values for each marker.}

\item{nb_bootstrap}{numeric, number of bootstrap (by defaults nb_bootstrap = 1000).}
}
\value{
list of 3 slots:
    - model_fit: fit of the model;
    - plot: plot of the effects;
    - results_summary: data.frame containing the results of the models for
    each marker.
}
\description{
Run cytoglm.
}
\details{
Function to run the Generalized Linear Model with Bootstrap,
from the CytoGLMM package.
}
