% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{function_run_cytoGLMM}
\alias{function_run_cytoGLMM}
\title{Run cytoglmm}
\usage{
function_run_cytoGLMM(mock_dataset)
}
\arguments{
\item{mock_dataset}{data.frame, cell values for each marker.}
}
\value{
list of 3 slots:
    - model_fit: fit of the model;
    - plot: plot of the effects;
    - results_summary: data.frame containing the results of the models for
    each marker.
}
\description{
Run cytoglmm
}
\details{
Function to run GLMM model from CytoGLMM package.
}
\keyword{internal}
