% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{function_run_diffcyt_full_pipeline}
\alias{function_run_diffcyt_full_pipeline}
\title{Run diffcyt pipeline}
\usage{
function_run_diffcyt_full_pipeline(
  onevariation,
  model = c("limma", "LMM"),
  effect = c("random", "fixed")
)
}
\arguments{
\item{onevariation}{list, of simulated data (output of the
function_wrapper_apply_simulation_nbtimes function).}

\item{model}{character, model to run: "limma" or "LMM".}

\item{effect}{character, effect: "random" or "fixed".}
}
\value{
list with 2 slots:
    - model_fit: the model fit;
    - result_summary: results of the model for each marker.
}
\description{
Run diffcyt pipeline
}
\details{
Function to run diffcyt pipeline. We do not used the diffcyt function
directly because of the limma model run differently with the effects.
}
\keyword{internal}
