% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{function_run_diffcytDSlimma}
\alias{function_run_diffcytDSlimma}
\title{Run diffcyt-DS-limma model.}
\usage{
function_run_diffcytDSlimma(
  ls_desigmat_contrast,
  df_experiment_info,
  ls_features
)
}
\arguments{
\item{ls_desigmat_contrast}{list, with design and contrast matrices.}

\item{df_experiment_info}{data.frame, information about the experiment (donor IDs,
group IDs and sample IDs).}

\item{ls_features}{list, with cell counts and medians for each markers.}
}
\value{
list with 2 slots:
    - model_fit: the model fit;
    - result_summary: results of the model for each marker.
}
\description{
Run diffcyt-DS-limma model.
}
\details{
Function to run diffcyt-DS-limma model from the diffcyt package.
}
\keyword{internal}
