% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{function_run_diffcytDSLMM}
\alias{function_run_diffcytDSLMM}
\title{Run diffcyt-DS-LMM model.}
\usage{
function_run_diffcytDSLMM(ls_form_contrast, df_experiment_info, ls_features)
}
\arguments{
\item{ls_form_contrast}{list, with formula and constrat matrix.}

\item{df_experiment_info}{data.frame, information about the experiment (donor IDs,
group IDs and sample IDs).}

\item{ls_features}{list, with cell counts and medians for each markers.}
}
\value{
list with 2 slots:
    - model_fit: the model fit;
    - result_summary: results of the model for each marker.
}
\description{
Run diffcyt-DS-LMM model.
}
\details{
Function to run diffcyt-DS-LMM model from the diffcyt package.
}
