% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_model_results.R
\name{function_summary_results_models}
\alias{function_summary_results_models}
\title{Summarize data and results.}
\usage{
function_summary_results_models(summary_from_model, package)
}
\arguments{
\item{summary_from_model}{list, output from the functions running the models.}

\item{package}{character, package used to run the test: "CytoGLMM" or "diffcyt".}
}
\value{
data.frame of results for each simulation.
}
\description{
Summarize data and results.
}
\details{
Function to do a summary of the tested data and model's results
  for the CytoGLMM and diffcyt packages.
}
\keyword{internal}
