% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_nb_parameters_optimized.R
\name{function_to_transform_data}
\alias{function_to_transform_data}
\title{arcsinh transformation.}
\usage{
function_to_transform_data(data, cofactor = 5)
}
\arguments{
\item{data}{data.frame, cell values to transform.}

\item{cofactor}{numeric, co-factor used in the arcsinh (by default
cofactor = 5).}
}
\value{
data.frame of transformed data.
}
\description{
arcsinh transformation.
}
\details{
Function to transform the data with the recommended transformation
for cyTOF data: arcsinh with cofactor equals to 5.
}
